+++ openrem/remapp/tests/test_export_dx_xlsx.py
@@ -31,11 +31,15 @@ class ExportDXxlsx(TestCase):
         dx_ge_xr220_1 = os.path.join("test_files", "DX-Im-GE_XR220-1.dcm")
         dx_ge_xr220_2 = os.path.join("test_files", "DX-Im-GE_XR220-2.dcm")
         dx_ge_xr220_3 = os.path.join("test_files", "DX-Im-GE_XR220-3.dcm")
+        dx_carestream_dr7500_1 = os.path.join("test_files", "DX-Im-Carestream_DR7500-1.dcm")
+        dx_carestream_dr7500_2 = os.path.join("test_files", "DX-Im-Carestream_DR7500-2.dcm")
         root_tests = os.path.dirname(os.path.abspath(__file__))
 
         dx(os.path.join(root_tests, dx_ge_xr220_1))
         dx(os.path.join(root_tests, dx_ge_xr220_2))
         dx(os.path.join(root_tests, dx_ge_xr220_3))
+        dx(os.path.join(root_tests, dx_carestream_dr7500_1))
+        dx(os.path.join(root_tests, dx_carestream_dr7500_2))
 
     def test_id_as_text(self):  # See https://bitbucket.org/openrem/openrem/issues/443
         filter_set = ""
@@ -63,6 +67,35 @@ class ExportDXxlsx(TestCase):
         self.assertEqual(all_data_sheet.cell_value(1, patient_id_col), '00098765')
         self.assertEqual(all_data_sheet.cell_value(1, accession_number_col), '00938475')
         self.assertEqual(all_data_sheet.cell_value(1, exposure_index_col), 51.745061)
+
+        # cleanup
+        task.filename.delete()  # delete file so local testing doesn't get too messy!
+        task.delete()  # not necessary, by hey, why not?
+
+    def test_filters(self):
+        # Tests that extracts with multiple filters succeed (though previous test would fail too!
+        filter_set = ''
+        # filter_set = "display_name=Carestream+Clinic+KODAK7500&"
+        pid = True
+        name = True
+        patient_id = True
+
+        dxxlsx(filter_set, pid=pid, name=name, patid=patient_id, user=self.user)
+
+        import xlrd
+        task = Exports.objects.all()[0]
+
+        book = xlrd.open_workbook(task.filename.path)
+        aec_sheet = book.sheet_by_name('aec')
+        headers = aec_sheet.row(0)
+
+        filter_col = [i for i, x in enumerate(headers) if x.value == 'Filters'][0]
+        filter_thick_col = [i for i, x in enumerate(headers) if x.value == 'Filter thicknesses (mm)'][0]
+
+        self.assertEqual(aec_sheet.cell_value(1, filter_col), 'Al')
+        self.assertEqual(aec_sheet.cell_value(1, filter_thick_col), '1.0')
+        self.assertEqual(aec_sheet.cell_value(2, filter_col), 'Al | Cu')
+        self.assertEqual(aec_sheet.cell_value(2, filter_thick_col), '1.0 | 0.2')
 
         # cleanup
         task.filename.delete()  # delete file so local testing doesn't get too messy!